+++ lib/Polynomials/polynomials.h
@@ -6,8 +6,8 @@
 #include <BasicLinearAlgebra.h>
 #include "BLA_tools.h"
 
-double evaluatePolynomial(const std::vector<double> &coefficients, double t);
+float evaluatePolynomial(const std::vector<float> &coefficients, float t);
 
-std::vector<double> differentiatePolynomial(const std::vector<double> &coefficients);
+std::vector<float> differentiatePolynomial(const std::vector<float> &coefficients);
 
-std::vector<double> quintic_poly(const std::vector<double> &q1, const std::vector<double> &q2);
+std::vector<float> quintic_poly(const std::vector<float> &q1, const std::vector<float> &q2);
